Rename packages. Add GPS arrow.
