+++ android/java/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxgl/geometry/IProjectedMeters.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxgl.lib.geometry;
+package com.mapbox.mapboxgl.geometry;
 
 public interface IProjectedMeters {
     double getNorthing();
