+++ android/java/MapboxGLAndroidSDKTestApp/build.gradle
@@ -38,7 +38,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.mapbox.mapboxgl.app"
+        applicationId "com.mapbox.mapboxgl.testapp"
         minSdkVersion 14
         targetSdkVersion 21
         versionCode 4
